+++ newclient/src/modules/tokens/components/TableTokensHeader.vue
@@ -15,30 +15,30 @@
         =====================================================================================
         -->
         <v-flex sm4>
-            <v-layout align-center justify-start row pl-4>
+            <v-layout :class="[!isActive(0) && !isActive(1) ? 'inactive-sort' : '']" align-center justify-start row pl-4>
                 <h5 class="pl-5 pr-2">{{ $tc('token.name', 1) }}</h5>
-                <v-flex>
-                    <v-layout align-start justify-center column>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(0)">
-                            <v-icon :class="[isActive(0) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-up</v-icon>
+                <v-flex d-flex align-center>
+                    <v-layout align-start justify-right>
+                        <v-btn v-if="!loading && !isActive(1)" :class="[!isActive(0) && !isActive(1) ? 'inactive-btn' : '']" flat icon @click="selectFilter(0)">
+                            <v-icon :class="[isActive(0) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
                         </v-btn>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(1)">
-                            <v-icon :class="[isActive(1) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-down</v-icon>
+                        <v-btn v-if="!loading && isActive(1)" flat icon @click="selectFilter(1)">
+                            <v-icon :class="[isActive(1) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
                         </v-btn>
                     </v-layout>
                 </v-flex>
             </v-layout>
         </v-flex>
         <v-flex sm2>
-            <v-layout align-center justify-start row pl-1>
+            <v-layout :class="[!isActive(2) && !isActive(3) ? 'inactive-sort' : '']" align-center justify-start row pl-1>
                 <h5 class="pr-2">{{ $tc('price.name', 1) }}</h5>
-                <v-flex>
-                    <v-layout align-start justify-center column>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(2)">
-                            <v-icon :class="[isActive(2) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-up</v-icon>
+                <v-flex d-flex align-center>
+                    <v-layout align-start justify-right>
+                        <v-btn v-if="!loading && !isActive(3)" :class="[!isActive(2) && !isActive(3) ? 'inactive-btn' : '']" flat icon @click="selectFilter(2)">
+                            <v-icon :class="[isActive(2) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
                         </v-btn>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(3)">
-                            <v-icon :class="[isActive(3) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-down</v-icon>
+                        <v-btn v-if="!loading && isActive(3)" flat icon @click="selectFilter(3)">
+                            <v-icon :class="[isActive(3) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
                         </v-btn>
                     </v-layout>
                 </v-flex>
@@ -48,30 +48,30 @@
             <h5>{{ $t('token.change') }}</h5>
         </v-flex>
         <v-flex sm2>
-            <v-layout align-center justify-start row pl-2>
+            <v-layout :class="[!isActive(4) && !isActive(5) ? 'inactive-sort' : '']" align-center justify-start row pl-2>
                 <h5 class="pr-1">{{ $t('token.volume') }}</h5>
-                <v-flex>
-                    <v-layout align-start justify-center column>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(4)">
-                            <v-icon :class="[isActive(4) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-up</v-icon>
+                <v-flex d-flex align-center>
+                    <v-layout align-start justify-right>
+                        <v-btn v-if="!loading && !isActive(5)" :class="[!isActive(4) && !isActive(5) ? 'inactive-btn' : '']" flat icon @click="selectFilter(4)">
+                            <v-icon :class="[isActive(4) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
                         </v-btn>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(5)">
-                            <v-icon :class="[isActive(5) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-down</v-icon>
+                        <v-btn v-if="!loading && isActive(5)" flat icon @click="selectFilter(5)">
+                            <v-icon :class="[isActive(5) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
                         </v-btn>
                     </v-layout>
                 </v-flex>
             </v-layout>
         </v-flex>
         <v-flex sm2>
-            <v-layout align-center justify-start row pl-2>
+            <v-layout :class="[!isActive(6) && !isActive(7) ? 'inactive-sort' : '']" align-center justify-start row pl-2>
                 <h5 class="pr-1">{{ $t('token.market') }}</h5>
-                <v-flex>
-                    <v-layout align-start justify-center column>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(6)">
-                            <v-icon :class="[isActive(6) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-up</v-icon>
+                <v-flex d-flex align-center>
+                    <v-layout align-start justify-right>
+                        <v-btn v-if="!loading && !isActive(7)" :class="[!isActive(6) && !isActive(7) ? 'inactive-btn' : '']" flat icon @click="selectFilter(6)">
+                            <v-icon :class="[isActive(6) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
                         </v-btn>
-                        <v-btn :disabled="loading" flat icon @click="selectFilter(7)">
-                            <v-icon :class="[isActive(7) ? 'white--text' : 'bttnToken--text']" small>fas fa-caret-down</v-icon>
+                        <v-btn v-if="!loading && isActive(7)" flat icon @click="selectFilter(7)">
+                            <v-icon :class="[isActive(7) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
                         </v-btn>
                     </v-layout>
                 </v-flex>
@@ -115,6 +115,13 @@ export default class TableTokensHeader extends Vue {
      * @param _value {Number}
      */
     selectFilter(_value: number): void {
+        if (this.isActive(_value)) {
+            if (_value % 2 == 0) {
+                _value = _value + 1
+            } else {
+                _value = _value - 1
+            }
+        }
         this.$emit('sortBy', FILTER_VALUES[_value])
     }
     /**
@@ -138,10 +145,24 @@ export default class TableTokensHeader extends Vue {
 }
 </script>
 
-<style scoped lang="css">
+<style scoped lang="scss">
 .v-btn.v-btn--flat.v-btn--icon {
-    height: 12px;
-    width: 12px;
+    height: 16px;
+    width: 16px;
     margin: 0;
+    .v-icon {
+        font-size: 18px !important;
+    }
+}
+.inactive-sort {
+    .inactive-btn {
+        color: transparent;
+    }
+    &:hover {
+        .v-icon {
+            color: #b4bfd2;
+            display: block;
+        }
+    }
 }
 </style>